+++ vscode/src/completions/get-inline-completions.ts
@@ -10,8 +10,9 @@ import { GetContextOptions, GetContextResult } from './context/context'
 import { GraphContextFetcher } from './context/context-graph'
 import { DocumentHistory } from './context/history'
 import { DocumentContext } from './get-current-doc-context'
+import { AutocompleteItem } from './inline-completion-item-provider'
 import * as CompletionLogger from './logger'
-import { SuggestionID } from './logger'
+import { CompletionLogID } from './logger'
 import { CompletionProviderTracer, Provider, ProviderConfig, ProviderOptions } from './providers/provider'
 import { RequestManager, RequestParams } from './request-manager'
 import { reuseLastCandidate } from './reuse-last-candidate'
@@ -54,13 +55,10 @@ export interface InlineCompletionsParams {
 
     // Callbacks to accept completions
     handleDidAcceptCompletionItem?: (
-        logId: SuggestionID,
-        completion: InlineCompletionItemWithAnalytics,
-        request: RequestParams
+        completion: Pick<AutocompleteItem, 'requestParams' | 'logId' | 'analyticsItem' | 'trackedRange'>
     ) => void
     handleDidPartiallyAcceptCompletionItem?: (
-        logId: SuggestionID,
-        completion: InlineCompletionItemWithAnalytics,
+        completion: Pick<AutocompleteItem, 'logId' | 'analyticsItem'>,
         acceptedLength: number
     ) => void
 }
@@ -90,7 +88,7 @@ export interface LastInlineCompletionCandidate {
  */
 export interface InlineCompletionsResult {
     /** The unique identifier for logging this result. */
-    logId: SuggestionID
+    logId: CompletionLogID
 
     /** Where this result was generated from. */
     source: InlineCompletionsResultSource
@@ -122,13 +120,13 @@ export enum InlineCompletionsResultSource {
  * via the keyboard shortcut and invoking a completion via hovering over ghost text.
  */
 export enum TriggerKind {
-    /** Completion was triggered explicitly by a user hovering over ghost text. **/
+    /** Completion was triggered explicitly by a user hovering over ghost text. */
     Hover = 'Hover',
 
-    /** Completion was triggered automatically while editing. **/
+    /** Completion was triggered automatically while editing. */
     Automatic = 'Automatic',
 
-    /** Completion was triggered manually by the user invoking the keyboard shortcut. **/
+    /** Completion was triggered manually by the user invoking the keyboard shortcut. */
     Manual = 'Manual',
 
     /** When the user uses the suggest widget to cycle through different completions. */
@@ -228,7 +226,7 @@ async function doGetInlineCompletions(params: InlineCompletionsParams): Promise<
     // Only log a completion as started if it's either served from cache _or_ the debounce interval
     // has passed to ensure we don't log too many start events where we end up not doing any work at
     // all.
-    CompletionLogger.flushActiveSuggestions()
+    CompletionLogger.flushActiveSuggestionRequests()
     const multiline = Boolean(multilineTrigger)
     const logId = CompletionLogger.create({
         multiline,
@@ -305,7 +303,7 @@ async function doGetInlineCompletions(params: InlineCompletionsParams): Promise<
             ? InlineCompletionsResultSource.CacheAfterRequestStart
             : InlineCompletionsResultSource.Network
 
-    CompletionLogger.loaded(logId, reqContext, completions)
+    CompletionLogger.loaded(logId, reqContext, completions, source)
 
     return {
         logId,