+++ vscode/src/completions/tree-sitter/queries/python.ts
@@ -1,0 +1,14 @@
+import dedent from 'dedent'
+
+import { SupportedLanguage } from '../grammars'
+import type { QueryName } from '../queries'
+
+export const pythonQueries = {
+    [SupportedLanguage.Python]: {
+        blocks: dedent`
+            (_(":") @block_start) @trigger
+        `,
+        singlelineTriggers: '',
+        intents: '',
+    },
+} satisfies Partial<Record<SupportedLanguage, Record<QueryName, string>>>