+++ vscode/src/configuration.test.ts
@@ -31,6 +31,7 @@ describe('getConfiguration', () => {
             inlineChat: true,
             codeActions: true,
             isRunningInsideAgent: false,
+            agentIDE: undefined,
             experimentalNonStop: false,
             debugEnable: false,
             debugVerbose: false,
@@ -113,6 +114,8 @@ describe('getConfiguration', () => {
                         return 'lsp-light'
                     case 'cody.advanced.agent.running':
                         return false
+                    case 'cody.advanced.agent.ide':
+                        return undefined
                     default:
                         throw new Error(`unexpected key: ${key}`)
                 }
@@ -142,6 +145,7 @@ describe('getConfiguration', () => {
             inlineChat: true,
             codeActions: true,
             isRunningInsideAgent: false,
+            agentIde: undefined,
             experimentalNonStop: true,
             debugEnable: true,
             debugVerbose: true,