+++ vscode/CHANGELOG.md
@@ -12,6 +12,14 @@ Starting from `0.2.0`, Cody is using `major.EVEN_NUMBER.patch` for release versi
 
 ### Changed
 
+## [0.14.5]
+
+### Added
+
+### Fixed
+
+### Changed
+
 - Added support to test a Sourcegraph specific StarCoder setup for dotcom. [pull/1670]
 
 ## [0.14.4]